v2v Refactoring
